pyproject.toml
[project]
name = "daily-news"
version = "0.1.0"
description = "World news aggregator with AI-powered ranking"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "feedparser>=6.0.10",
    "deep-translator>=1.11.0",
    "anthropic>=0.39.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.8.0",
    "mypy>=1.8.0",
]

[project.scripts]
daily-news = "daily_news.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
